Rename mod.py -> mods.py, since sometimes we would like Mod objects to be called mod.
